Fix. Comparison between signed and unsigned ints.
In the previous commit, I submitted code which compared signed and unsigned
integers, raising a compiler warning. Cast vector::size() to an int to suppress
warnings.

+++ Collectors/MarkSweepCollector.cpp
@@ -114,7 +114,7 @@ void MarkSweepCollector::enqueueAllRoots() {
 		vector<Object*> roots;
 		for (i = 0; i < NUM_THREADS; i++) {
 			roots = myObjectContainer->getRoots(i);
-			for (j = 0; j < roots.size(); j++) {
+			for (j = 0; j < (int)roots.size(); j++) {
 				currentObj = roots[j];
 				if (currentObj && currentObj->getVisited() == 0) {
 					currentObj->setVisited(1);
